[tracing] refactor to use custom extension for trace id editing (#5420)
refactor to use custom extension for trace id editing
