+++ config/envoyconfig/extensions/pomerium_otel.proto
@@ -1,0 +1,19 @@
+syntax = "proto3";
+
+package pomerium.extensions;
+
+import "envoy/config/core/v3/extension.proto";
+import "envoy/config/core/v3/grpc_service.proto";
+import "envoy/config/core/v3/http_service.proto";
+import "udpa/annotations/migrate.proto";
+import "udpa/annotations/status.proto";
+
+option (udpa.annotations.file_status).package_version_status = ACTIVE;
+
+message OpenTelemetryConfig {
+  envoy.config.core.v3.GrpcService                   grpc_service       = 1 [(udpa.annotations.field_migrate).oneof_promotion = "otlp_exporter"];
+  envoy.config.core.v3.HttpService                   http_service       = 3 [(udpa.annotations.field_migrate).oneof_promotion = "otlp_exporter"];
+  string                                             service_name       = 2;
+  repeated envoy.config.core.v3.TypedExtensionConfig resource_detectors = 4;
+  envoy.config.core.v3.TypedExtensionConfig          sampler            = 5;
+}
